do not leak resolver in remoteservice
Summary:
avahi's dbus api expects us to control the live time of objects. to that
end we need to hold our wrapper and free it when we no longer need it.
this already has cleanup code for m_resolver, unfortunately the member was
never set.

Test Plan: builds

Reviewers: broulik

Reviewed By: broulik

Subscribers: kde-frameworks-devel

Tags: #frameworks

Differential Revision: https://phabricator.kde.org/D17067